move NullDNSWriter to dns package
