+++ pkg/roles/dns/utils.go
@@ -1,4 +1,4 @@
-package dns_test
+package dns
 
 import (
 	"errors"
